fix(newtypes): Fix pydantic conversions of newtypes within a list
